+++ Runtime/DataStructs/Implementation/Builders/BimParentNodeBuilder.cs
@@ -1,0 +1,26 @@
+﻿// Copyright (c) 2023  DigitalTwin Technology GmbH
+// https://www.digitaltwin.technology/
+
+using UnityEngine;
+using DTTUnityCore.DataStructs;
+
+namespace DTTBim.DataStructs
+{
+    public class BimParentNodeBuilder : IDataNodeBuilder
+    {
+        private MetaDataName _metaDataName;
+
+        public IMetaData MetaData { get => _metaDataName; set => _metaDataName = (MetaDataName)value; }
+
+        public BimParentNodeBuilder(string name)
+        {
+            _metaDataName = new MetaDataName(name);
+        }
+
+        public IMetaDataNode<NodeType, MetaDataType> Create<NodeType, MetaDataType>(IMetaDataNode<NodeType, MetaDataType> parent) where MetaDataType : IMetaData
+        {
+            return (IMetaDataNode<NodeType, MetaDataType>)(new GameObject(_metaDataName.Name)).AddComponent<BimNodeParent>();
+        }
+    }
+}
+