+++ CHANGELOG.md
@@ -1,2 +1,25 @@
-# DTTUnityCommon
-Common Unity Libraries
+# DTTBim | Changelog 	 
+```
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+```
+## [Unreleased]
+## [0.1.1] - 07.07.2023.
+### Added
+1. Add first changelog md file
+2. Add Interface DTTBim.DataStructs.IBimParentNode
+3. Add Interface DTTBim.DataStructs.IBimNodeScene
+4. Add Interface DTTBim.DataStructs.IBimNodeObject
+5. Add class DTTBim.DataStructs.BimParentNode
+6. Add class DTTBim.DataStructs.BimNodeScene
+7. Add class DTTBim.DataStructs.BimNodeObject
+8. Add class DTTBim.DataStructs.BimNodeObjectBuilder
+9. Add class BimNodeSceneBuilder
+10. Add class BimParentNodeBuilder
+
+
+
+
+